Added a despair!(with($err), because($reason)) macro, along with a default handler that just loops.  The OS can potentially recover from a panic, but despair is intended for those scenarios that have no recovery, such as when we detect that the linker script has produced invalid code.
The despair handler, __smeg_is_in_despair, can be overridden to do something other than an infinite loop.  This could be useful, for example, on certain boards where you can provide a hand-rolled assembly-language stub to output an SOS pattern to a GPIO or UART.  The handler is deeply unsafe and nothing can be assumed, since we probably ended up in despair due to invalid code generation.