Added two (feature-toggled) implementations for the BssSectionInitialiser trait, one ensuring the linker script gives sane(ish) symbols, the other skipping the checks.
This work also resulted in factoring the integration-testing boilerplate for despair!(...) into a new crate, and also highlighted the issue for documentation due to the unsafe contracts on the various traits.  Documentation wants to be in side-by-side markdown rather than inline with the code (too much clutter), so there are some TODOs for implementing a #[link_doc] attribute.