Started to think about and implement the Syscall mechanism.  For the Cortex M4 devices (and plenty of other ARMs) this will be the SV_CALL interrupt.  Added various implementation notes for TODO, but committed as a WIP to save current progress.
The Syscall ISR will be implemented as a Driver, so during bootstrapping the kernel will simply handle this in the same way as other Drivers.  All the other Drivers can then build off the Syscall implementation, which means we ought to be able to drop out of privileged execution quite soon, since privilege escalation can then be accomplished by the Syscalls for just the parts of Driver initialisation that require it, although it won't be full protection because swathes of infrastructure (eg. any MMU) will still be uninitialised it's better than nothing.