Started to work the Syscall PoC into the codebase.  The mechanism is quite fragile due to some limitations with Rust const contexts and generic statics, but it's workable.
The implementation depends on linker script symbols and 'magic' sections / names, but this approach can be made to work in a manner that allows each Driver to create their own Syscalls, allowing for an extensible OS API without having one or more master lists of Syscalls (ie. the Open-Closed Principle).

WIP intermediate commit because there are a lot of moving parts here but it compiles (correctly) so it needs to be banked.